% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_plot.R
\name{aa_plot}
\alias{aa_plot}
\title{Plot number of amino acids in peptide}
\usage{
aa_plot(peptide)
}
\arguments{
\item{peptide}{}
}
\value{
Graph of amino acid counts
}
\description{
Plot number of amino acids in peptide
}
