% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_start.R
\name{codon_start}
\alias{codon_start}
\title{Title: Start codon_RNA sequence to aminoacid}
\usage{
codon_start(rna, start = 1)
}
\arguments{
\item{rna}{sequence}

\item{start}{start of codon sequence}
}
\value{
codon
}
\description{
Title: Start codon_RNA sequence to aminoacid
}
