% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_table.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Codon Table}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_table
}
\description{
Codon Table for tranlsation of RNA to Protein
}
\keyword{datasets}
