% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_codon.R
\name{start_codon}
\alias{start_codon}
\title{Title: Start codon_RNA sequence to aminoacid}
\usage{
start_codon(rna, start = 1)
}
\arguments{
\item{rna}{sequence}

\item{start}{start codon}
}
\value{
codon
}
\description{
Title: Start codon_RNA sequence to aminoacid
}
