% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_one.R
\name{sample_with_replacement}
\alias{sample_with_replacement}
\title{Sample the nucleotides}
\usage{
sample_with_replacement(nucleotide_size)
}
\arguments{
\item{nucleotide_size}{}
}
\value{

}
\description{
Sample the nucleotides
}
